requirements.txt
fastapi
uvicorn 
sqlalchemy
python-jose[cryptography]
#passlib[bcrypt]
python-dotenv
bcrypt
